refactor: split tools/mod.rs into category modules (P9)
Split the 4948-line monolith into 14 focused files using the delegate
pattern. Each #[tool] method in the tool_router impl block is now a
thin stub that delegates to a handle_* function in its category module.

- helpers.rs: cross-client filter, compact mode, JSON result helpers
- shared.rs: embed_and_store, get_query_embedding, build_client_lookup
- Category modules (inventory, runbooks, knowledge, context, incidents,
  coordination, search, monitoring, zammad, briefings): param structs +
  handler implementations colocated
- OpsBrain fields changed to pub(crate) for category module access
- mod.rs reduced from 4948 to 1183 lines (76% reduction)
- CLAUDE.md updated with new project layout and contributing guide

Also includes P4 (push monitor diagnostic hints): push-type monitors
that are DOWN now include a diagnostic_hint field explaining that
"down" usually means heartbeat expired, not service failure.

All 101 tests pass (81 unit + 20 integration), clippy clean.
